+++ next-docs/public/examples/avatar/Variants.tsx
@@ -1,9 +1,12 @@
 import { Avatar } from '@heathmont/moon-core-tw';
+import { GenericUser } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
   <>
-    <Avatar />
-    <Avatar name="md" />
+    <Avatar>
+      <GenericUser className="text-moon-24" />
+    </Avatar>
+    <Avatar>md</Avatar>
     <Avatar imageUrl="../avatar.jpeg" />
   </>
 );