+++ workspaces/core/src/accordion/private/utils/setFont.ts
@@ -4,6 +4,8 @@ const setFont = (size?: AccordionSizes) => {
   switch (size) {
     case 'xl':
       return `text-moon-16`;
+    case 'sm':
+      return `text-moon-12`;
     default:
       return `text-moon-14`;
   }