+++ workspaces/core/src/avatar/private/utils/setStatusSize.ts
@@ -2,12 +2,12 @@ import type SizeProps from '../types/SizeProps';
 
 const setStatusSize = (size?: SizeProps) => {
   if (size === 'xs') {
-    return 'w-2 h-2 border';
+    return '[&_.status]:w-2 [&_.status]:h-2 [&_.status]:border';
   }
   if (size === 'sm' || size === 'md') {
-    return 'w-3 h-3 border-2';
+    return '[&_.status]:w-3 [&_.status]:h-3 [&_.status]:border-2';
   }
-  return 'w-4 h-4 border-2';
+  return '[&_.status]:w-4 [&_.status]:h-4 [&_.status]:border-2';
 };
 
 export default setStatusSize;